Create Lab12
So I changed up the code to no longer use the tryAgain function. Instead, if you chose the wrong choice (not in the list of commands) you are prompted with the same question again. The problem is that when this runs and you exit, it gets stuck in a loop bouncing back and forth between the move function and whatever room it is on. Check it out